+++ lib/semantic_logger.rb
@@ -1,51 +1,23 @@
+require "concurrent"
 require "semantic_logger/core_ext/thread"
 require "semantic_logger/version"
-
-# @formatter:off
-module SemanticLogger
-  autoload :AnsiColors,         "semantic_logger/ansi_colors"
-  autoload :Appender,           "semantic_logger/appender"
-  autoload :Appenders,          "semantic_logger/appenders"
-  autoload :Base,               "semantic_logger/base"
-  autoload :DebugAsTraceLogger, "semantic_logger/debug_as_trace_logger"
-  autoload :Formatters,         "semantic_logger/formatters"
-  autoload :Levels,             "semantic_logger/levels"
-  autoload :Log,                "semantic_logger/log"
-  autoload :Logger,             "semantic_logger/logger"
-  autoload :Loggable,           "semantic_logger/loggable"
-  autoload :Processor,          "semantic_logger/processor"
-  autoload :QueueProcessor,     "semantic_logger/queue_processor"
-  autoload :Subscriber,         "semantic_logger/subscriber"
-  autoload :SyncProcessor,      "semantic_logger/sync_processor"
-  autoload :Utils,              "semantic_logger/utils"
-
-  module Concerns
-    autoload :Compatibility, "semantic_logger/concerns/compatibility"
-  end
-
-  module Metric
-    autoload :NewRelic,          "semantic_logger/metric/new_relic"
-    autoload :Signalfx,          "semantic_logger/metric/signalfx"
-    autoload :Statsd,            "semantic_logger/metric/statsd"
-  end
-
-  module Reporters
-    autoload :Minitest,          "semantic_logger/reporters/minitest"
-  end
-
-  module Test
-    autoload :CaptureLogEvents,  "semantic_logger/test/capture_log_events"
-    autoload :Minitest,          "semantic_logger/test/minitest"
-  end
-
-  if defined?(JRuby)
-    module JRuby
-      autoload :GarbageCollectionLogger, "semantic_logger/jruby/garbage_collection_logger"
-    end
-  end
-end
+require "semantic_logger/utils"
+require "semantic_logger/ansi_colors"
+require "semantic_logger/levels"
+require "semantic_logger/base"
+require "semantic_logger/formatters"
+require "semantic_logger/log"
+require "semantic_logger/subscriber"
+require "semantic_logger/loggable"
+require "semantic_logger/concerns/compatibility"
+require "semantic_logger/appender"
+require "semantic_logger/appenders"
+require "semantic_logger/processor"
+require "semantic_logger/queue_processor"
+require "semantic_logger/sync_processor"
+require "semantic_logger/logger"
+require "semantic_logger/debug_as_trace_logger"
 require "semantic_logger/semantic_logger"
-# @formatter:on
 
 # Flush all appenders at exit, waiting for outstanding messages on the queue
 # to be written first.