+++ lib/semantic_logger/appender.rb
@@ -15,6 +15,7 @@ module SemanticLogger
     autoload :Http,              "semantic_logger/appender/http"
     autoload :MongoDB,           "semantic_logger/appender/mongodb"
     autoload :NewRelic,          "semantic_logger/appender/new_relic"
+    autoload :NewRelicLogs,      "semantic_logger/appender/new_relic_logs"
     autoload :Rabbitmq,          "semantic_logger/appender/rabbitmq"
     autoload :Splunk,            "semantic_logger/appender/splunk"
     autoload :SplunkHttp,        "semantic_logger/appender/splunk_http"