+++ lib/semantic_logger/formatters/base.rb
@@ -23,7 +23,6 @@ module SemanticLogger
       #     See Time#strftime for the format of this string.
       #     :iso_8601 Outputs an ISO8601 Formatted timestamp.
       #     :ms       Output in miliseconds since epoch.
-      #     nil:      Returns Empty string for time ( no time is output ).
       #     Default: '%Y-%m-%d %H:%M:%S.%<precision>N'
       #   log_host: [Boolean]
       #     Whether or not to include hostname in logs