+++ lib/semantic_logger/log.rb
@@ -219,11 +219,11 @@ module SemanticLogger
 
       seconds = duration / 1000
       if seconds >= 86_400.0 # 1 day
-        "#{(seconds / 86_400).to_i}d #{Time.at(seconds).strftime('%-Hh %-Mm')}"
+        "#{(seconds / 86_400).to_i}d #{Time.at(seconds).utc.strftime('%-Hh %-Mm')}"
       elsif seconds >= 3600.0 # 1 hour
-        Time.at(seconds).strftime("%-Hh %-Mm")
+        Time.at(seconds).utc.strftime("%-Hh %-Mm")
       elsif seconds >= 60.0 # 1 minute
-        Time.at(seconds).strftime("%-Mm %-Ss")
+        Time.at(seconds).utc.strftime("%-Mm %-Ss")
       elsif seconds >= 1.0 # 1 second
         "#{format('%.3f', seconds)}s"
       else
@@ -248,7 +248,11 @@ module SemanticLogger
 
     # Extract the filename and line number from the last entry in the supplied backtrace
     def extract_file_and_line(stack, short_name = false)
+      return unless stack&.size&.positive?
+
       match = CALLER_REGEXP.match(stack.first)
+      return unless match
+
       [short_name ? File.basename(match[1]) : match[1], match[2].to_i]
     end
 
@@ -256,7 +260,7 @@ module SemanticLogger
     # in either the backtrace or exception
     def file_name_and_line(short_name = false)
       stack = backtrace || exception&.backtrace
-      extract_file_and_line(stack, short_name) if stack&.size&.positive?
+      extract_file_and_line(stack, short_name)
     end
 
     # Strip the standard Rails colorizing from the logged message