+++ test/appender/http_test.rb
@@ -3,33 +3,37 @@ require_relative "../test_helper"
 # Unit Test for SemanticLogger::Appender::Http
 module Appender
   class HttpTest < Minitest::Test
-    response_mock = Struct.new(:code, :body)
-
     describe SemanticLogger::Appender::Http do
-      let(:appender) { SemanticLogger::Appender::Http.new(url: "http://mockhost:8088/path") }
-      let(:amessage) { "AppenderHttpTest log message" }
+      let(:http_success) { Net::HTTPSuccess.new("1.1", "200", "OK") }
+      let(:log_message) { "AppenderHttpTest log message" }
+
+      let(:appender) do
+        Net::HTTP.stub_any_instance(:start, true) do
+          SemanticLogger::Appender::Http.new(url: "http://localhost:8088/path")
+        end
+      end
 
       SemanticLogger::LEVELS.each do |level|
         it "send #{level}" do
           request = nil
-          appender.http.stub(:request, ->(r) { request = r; response_mock.new("200", "ok") }) do
-            appender.send(level, amessage)
+          appender.http.stub(:request, ->(r) { request = r; http_success }) do
+            appender.send(level, log_message)
           end
           hash = JSON.parse(request.body)
-          assert_equal amessage, hash["message"]
+          assert_equal log_message, hash["message"]
           assert_equal level.to_s, hash["level"]
           refute hash["stack_trace"]
         end
 
         it "send #{level} exceptions" do
-          exc     = nil
+          exc = nil
           begin
             Uh oh
           rescue Exception => e
             exc = e
           end
           request = nil
-          appender.http.stub(:request, ->(r) { request = r; response_mock.new("200", "ok") }) do
+          appender.http.stub(:request, ->(r) { request = r; http_success }) do
             appender.send(level, "Reading File", exc)
           end
           hash = JSON.parse(request.body)
@@ -42,16 +46,38 @@ module Appender
 
         it "send #{level} custom attributes" do
           request = nil
-          appender.http.stub(:request, ->(r) { request = r; response_mock.new("200", "ok") }) do
-            appender.send(level, amessage, key1: 1, key2: "a")
+          appender.http.stub(:request, ->(r) { request = r; http_success }) do
+            appender.send(level, log_message, key1: 1, key2: "a")
           end
           hash = JSON.parse(request.body)
-          assert_equal amessage, hash["message"]
+          assert_equal log_message, hash["message"]
           assert_equal level.to_s, hash["level"]
           refute hash["stack_trace"]
           assert payload = hash["payload"], hash
           assert_equal 1, payload["key1"], payload
           assert_equal "a", payload["key2"], payload
+        end
+      end
+
+      it "supports http 204 success" do
+        http_success = Net::HTTPSuccess.new("1.1", "204", "OK")
+        request = nil
+        appender.http.stub(:request, ->(r) { request = r; http_success }) do
+          appender.info(log_message)
+        end
+        hash = JSON.parse(request.body)
+        assert_equal log_message, hash["message"]
+      end
+
+      it "supports custom headers" do
+        Net::HTTP.stub_any_instance(:start, true) do
+          request = nil
+          header = {"Authorization" => "Bearer BEARER_TOKEN"}
+          appender = SemanticLogger::Appender::Http.new(url: "http://localhost:8088/path", header: header)
+          appender.http.stub(:request, ->(r) { request = r; http_success }) do
+            appender.info(log_message)
+          end
+          assert_equal(header["Authorization"], request["Authorization"])
         end
       end
 
@@ -63,43 +89,49 @@ module Appender
       # Unfortunately this probably also means that this test will fail if it's
       # run on a machine that cannot resolve hostnames
       it "uses a proxy if specified" do
-        proxy    = "http://user:password@proxy.example.com:12345"
-        appender = SemanticLogger::Appender::Http.new(url: "http://mockhost:8088/path", proxy_url: proxy)
+        proxy = "http://user:password@proxy.example.com:12345"
+        Net::HTTP.stub_any_instance(:start, true) do
+          appender = SemanticLogger::Appender::Http.new(url: "http://ruby-lang.org:8088/path", proxy_url: proxy)
 
-        proxy_uri = URI.parse(proxy)
-        assert(appender.http.proxy?)
-        refute(appender.http.proxy_from_env?)
-        assert_equal(proxy_uri.host, appender.http.proxy_address)
-        assert_equal(proxy_uri.port, appender.http.proxy_port)
-        assert_equal(proxy_uri.user, appender.http.proxy_user)
-        assert_equal(proxy_uri.password, appender.http.proxy_pass)
+          proxy_uri = URI.parse(proxy)
+          assert(appender.http.proxy?)
+          refute(appender.http.proxy_from_env?)
+          assert_equal(proxy_uri.host, appender.http.proxy_address)
+          assert_equal(proxy_uri.port, appender.http.proxy_port)
+          assert_equal(proxy_uri.user, appender.http.proxy_user)
+          assert_equal(proxy_uri.password, appender.http.proxy_pass)
+        end
       end
 
       it "uses the ENV proxy if specified" do
-        old_env_proxy     = ENV["http_proxy"]
+        old_env_proxy = ENV.fetch("http_proxy", nil)
         ENV["http_proxy"] = "http://user:password@proxy.example.com:12345"
-        appender          = SemanticLogger::Appender::Http.new(url: "http://mockhost:8088/path")
+        Net::HTTP.stub_any_instance(:start, true) do
+          appender = SemanticLogger::Appender::Http.new(url: "http://ruby-lang.org:8088/path")
 
-        proxy_uri = URI.parse(ENV["http_proxy"])
-        assert(appender.http.proxy?)
-        assert(appender.http.proxy_from_env?)
-        assert_equal(proxy_uri.host, appender.http.proxy_address)
-        assert_equal(proxy_uri.port, appender.http.proxy_port)
-        assert_equal(proxy_uri.user, appender.http.proxy_user)
-        assert_equal(proxy_uri.password, appender.http.proxy_pass)
+          proxy_uri = URI.parse(ENV.fetch("http_proxy", nil))
+          assert(appender.http.proxy?)
+          assert(appender.http.proxy_from_env?)
+          assert_equal(proxy_uri.host, appender.http.proxy_address)
+          assert_equal(proxy_uri.port, appender.http.proxy_port)
+          assert_equal(proxy_uri.user, appender.http.proxy_user)
+          assert_equal(proxy_uri.password, appender.http.proxy_pass)
+        end
 
         ENV["http_proxy"] = old_env_proxy if old_env_proxy
       end
 
       it "doesn't use the ENV proxy if explicity requested" do
-        old_env_proxy     = ENV["http_proxy"]
+        old_env_proxy = ENV.fetch("http_proxy", nil)
         ENV["http_proxy"] = "http://user:password@proxy.example.com:12345"
-        appender          = SemanticLogger::Appender::Http.new(url: "http://mockhost:8088/path", proxy_url: nil)
+        Net::HTTP.stub_any_instance(:start, true) do
+          appender = SemanticLogger::Appender::Http.new(url: "http://ruby-lang.org:8088/path", proxy_url: nil)
 
-        refute(appender.http.proxy_from_env?)
-        refute(appender.http.proxy_address)
-        refute(appender.http.proxy_user)
-        refute(appender.http.proxy_pass)
+          refute(appender.http.proxy_from_env?)
+          refute(appender.http.proxy_address)
+          refute(appender.http.proxy_user)
+          refute(appender.http.proxy_pass)
+        end
 
         ENV["http_proxy"] = old_env_proxy if old_env_proxy
       end