+++ test/test_helper.rb
@@ -38,3 +38,7 @@ reporters = [
   SemanticLogger::Reporters::Minitest.new
 ]
 Minitest::Reporters.use!(reporters)
+
+def add_mocks_to_load_path
+  $LOAD_PATH.unshift File.join(File.dirname(__FILE__), "mocks")
+end